+++ packages/node/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Bring back `long` to fix missing filter `codeId` in long type
+
 ## [3.11.1] - 2024-05-02
 ### Fixed
 - Sandbox Uint8Array and missing pg dep issue